Mark non-notifying properties in event types CONSTANT
Event objects are valid only within the context of a JS callback that
receives the object as a signal parameter, and the object will not
change during the runtime of that callback.  It's not true that the
properties will never change, because event objects are pooled and
reused; but the QML is not allowed to keep a reference to an event,
so the property changes during this reuse are not meant to be seen.
QQuickMouseEvent is similar, and its properties were already marked
CONSTANT.

Change-Id: I4985e4a27bf192779629311eb8c6c1f63d22df3e
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>