QQmlTableInstanceModel: remove false assert
There current assert turned out to not be valid for
all cases. If the model is currently incubating an
object, m_modelItems list will contain it, but
the view will not. So be more specific what we
check. Also, we need to release the object for
items that are being incubated. The code snippet
for doing that is more or less the same as found
in the destructor of QQmlDelegateModel.

Change-Id: I84b4286a037b27ad809c3a63afed94ce61febc19
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>