Fix ASAN warnings about new-delete-size-mismatch
The optimization of allocating memory for a QML type along with its
declarative data in one shot works well and has worked well for many
years. However ASAN complains about it, because when the delete
expression is called from _anywhere_ on for example a QQuickItem, the
size returned by the virtual destructor and then passed to operator
delete is less than what was allocated earlier (to include the
declarative data structure).

Fix this by providing an operator delete for the final QQmlElement class
that we _do_ allocate, which is passed along by the virtual destructor.
That operator ignores the size and thus silences ASAN.

Task-number: QTBUG-58524
Change-Id: Ibb6120e7e07301bb4602130064c5157b21ec2202
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>