QQuickTableView: remove cacheBuffer from the implementation
After we removed cacheBuffer from the public API, giving
the users no way to switch if off, the safest thing is to
also remove it from the implementation. The cache buffer
can easily load add a lot of hidden items to the view, and
the user now has no way to tweak or hinder it.

As an example, lets say that you in a ListView can fit 10
items on screen. And then you have a cache buffer set that
loads two more items, both on top and below. You then end up with
14 items added to the view. Now, lets consider the same case
for TableView, where you show 10x10 items on screen. With the
same cache buffer, you end up loading 2x10 items in the
background on all sides of the table (pluss 4 items in each
corners). This sums up to 96 extra items. This is really bad
and unacceptable. It's more performant to just switch the
caching off completely.

Change-Id: Iddbd78ef1d7c7197eb4a847ec5067184149fe9a0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>