Fix explicit export of imported variables
Instead of using a re-export, it's also possible to write

    import { foo } from "./bar.js"

and then export it again

    export { foo }

Typically exported variables are referenced from the locals, but since
we don't add imports to the locals, we need another way of locating
them. This patch uses the index space after the locals in the internal
class for imports, so that after we've identifier the export in the
local export entry table, we can use the local name to search in the
internal class and find imports past the locals.

Change-Id: I58ab79ad3df1bbc1b972f0a2771d9ca1268de27b
Reviewed-by: Lars Knoll <lars.knoll@qt.io>