QQuickTableView: check if the providers are undefined, not null
Checking if QJSValue isNull() will just check if it contains
the js value "null". But we want to check if the application
has assigned anything at all to the providers.
Any value other than a function will not be accepted (which
we check for at the time we try to call them (from
resolveColumnWidth() and resolveColumnHeight()).

Change-Id: I24717b67e99dd1ad6684a83125d2a4c7826dd501
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>