Fix function name bindings for function declarations
For function expressions such as

    var foo = function foo() {
        return foo;
    }

it is important to have a separate binding for "foo" within the scope of
the function. However for function declarations this does not apply:

    function foo() {
        foo = 2;
    }
    foo()
    console.log(foo) // should print 2

Therefore we should enter the ThisFunctionName type of binding only for
function expressions or (generally) for the case where the name is not
intended to be entered into the surrounding environment at
defineFunction() time.

This is covered implicitly in

    language/module-code/eval-gtbndng-indirect-update-dflt.js

Change-Id: I1e5114a93ac7db9e5fcea04b1b3e1de4ad7bff6d
Reviewed-by: Lars Knoll <lars.knoll@qt.io>