Fix property declarations
It does not make sense to declare a property that has a setter and mark
it as constant. Moc also complains about it and notes how the constant
attribute is ignored. Therefore we might as well remove it from the
declaration.

Amends a9b9875acc6155f85558bb27cdb75f2dda5e2b66

Change-Id: Ib78d6e9c5e3415bc2a3dc26477d1d2a4e5b4dcfe
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>