Provide option to skip registration of enum classes unscoped
Enums classes are registered unscoped which leads to clashes in the following
cases:

1. Two different enums contain the same values
2. The name of an enum class is the same as an enum value

In the 2nd case you can not even access the scoped enum at all because it
will be overwritten by the primitive type.

Users can now add a class info to the meta type to disable the unscoped
registration which solves all clashes. The default is kept as is.

class MyClass : public QObject
{
    Q_OBJECT
    Q_CLASSINFO("RegisterEnumClassesUnscoped", "false")

public:
    ...
};

[ChangeLog][QtQml] Added option to disable unscoped registration of enum classes.

Change-Id: Ifa4197a14a252575e8a25ae56fb6ee479addf80b
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>