QQuickTableView: rename to scheduleRebuildTable, and add rebuild options
Rebuilding the table from scratch whenever e.g the model adds a new
row or column is slow and unnecessary. What happens is that we always
rebuild the table from the origin, and continue load and unload edges
until the loaded rows and columns overlaps with the viewport. This
can be slow if you are e.g at row 1000 when you start to rebuild.
Instead we can just start from current position in the viewport.

So add some options to control what needs to be done.

Note: This patch doesn't change any logic as it stands. But the options
will be used in a subsequent patch.

Change-Id: I9705dbae3a2c04e7e7189ec453756358a1b9fc14
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>