Fix enum warnings
QPlatformDialogHelper's meta-object is added to QQuickDialogButtonBox's
meta-object as a related meta object, and all of its enums are merged
into the same namespace by the QML engine. This produces a conflict with
the enum values of the ButtonLayout in QQuickDialogButtonBox, which is
a duplicate of the one that's already pulled in.

Fixes: QTBUG-70141
Task-number: QTBUG-70141
Change-Id: Ib33dc8ddbe8aa80d03183eb23861658c9e978f04
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>