Rename QQEventPoint::GrabState to GrabTransition
This enum represents a transient state transition, and only sometimes
corresponds to the current grab state of an event point.  For example
after exclusive grab has been canceled, the current state is that
there is no exclusive grab: it doesn't make sense to remember that the
way it got there was by cancellation.  There was an idea to add a
grabState property, but not all values would be eligible.  An
EventPoint can be exclusively grabbed by one item or handler at a
time, and by multiple passive grabbers at the same time, so even a
Q_FLAG would not fully express all possible states.  Besides, there is
already an exclusiveGrabber property, and we could add a
passiveGrabbers list property if we had a real need.  So adding a
grabState property seems unlikely, and therefore is not a good enough
reason to keep this enum named as GrabState.

Change-Id: Ie37742b4bd431a7e51910d79a7223fba9a6bd848
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>