QQuickTableView: check if the focusObject is a actually a QQuickItem
We used to do "static_cast<QQuickItem *>(window->focusObject());". But
window->focusObject() returns a QObject, which doesn't have to be a
QQuickItem. And in fact, during start-up, focusObject() will point to
the window itself. The result is that we could sometimes get a crash
because of this.

This patch will change the implementation to use qobject_cast instead.

Change-Id: Id5335a8efb4b2d400e308bf6c27158a406fc781e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>