QQuickTableView: be more precise about when to call updatePolish()
If a rebuild is scheduled, calling updatePolish() directly (which
is just an optimization) will rebuild the table there and then. But
we don't want this to happen upon a callback from viewportMoved(),
since rebuilding the table will usually also change the geometry of
the viewport/contentItem, which can easily trigger binding loops
back to the place that made the viewport move in the first place.

At the same time, we don't want to impose the same limitation when
calling updatePolish() from forceLayout(), since that function is
always called explicit from the application. And any binding loops
caused by it can as such be avoided by the developer.

This patch will therefore remove the common updatePolishIfPossible()
function, and move the logic for when to call updatePolish back
to the calling locations.

Change-Id: I845b9469c61735d27fc28ebe9f7b254d5a9b2efd
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>