Qt Labs Settings: warn if failed to initialize internal QSettings instance
First, warn the user that the QSettings instance failed to initialize
if there was an error. Then, if the error was QSettings::AccessError,
print the specific application identifiers that weren't set so they
know what they have to do. For example:

    QML Settings: Failed to initialize QSettings instance. Status code is: 1
    QML Settings: The following application identifiers have not been set: QVector("organizationName", "organizationDomain")

In my case, the QSettings instance wasn't created on
Windows because organizationName wasn't set.

Change-Id: I7970209e09b78f785730422c45978775311b96ac
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>